+++ setup.py
@@ -26,7 +26,7 @@ setup(name=about["__title__"],
       author=about["__author__"],
       author_email=about["__email__"],
       url=about["__uri__"],
-      packages=set(["aw_core", "aw_datastore", "aw_datastore.storages"]),
+      packages=set(["aw_core", "aw_transform", "aw_datastore", "aw_datastore.storages"]),
       package_data={"aw_core": ['schemas/*.json']},
       install_requires=install_requires,
       classifiers=[
